Stack-based Matrix Type, Complex Variable, Quaternion, Matrix using Complex Var
Added in new Matrix Storage that uses stack memory for more efficient
allocation and deallocation.

Built Complex variable type that can be used with the Matrix class.
Still need to add in functionality to be able to print both double and
complex variable types to the screen. Also need to add in the proper
complex transpose functionality.

Since the Stack-based Matrix type was available, updated the Quaternion
class to use this. Vec3D is a typedef of a 3x1 Stack-based Matrix.