Updated Fraction, Timer, Quaternion, vec3, FastMat
Fraction was updated such that it could accurately convert a
float/double into a fraction.

Timer was updated to put implementation in Timer.cpp file and leave
prototypes in Timer.hpp.

Quaternion class was updated with modified name convention for vec3,
which represents a 3D vector based on the FastMatrix class.